+++ test/server/index.ts
@@ -3,4 +3,7 @@ import * as admin from 'firebase-admin'
 
 admin.initializeApp()
 
-export const backupfire = backupfireAgent()
+export const backupfire = backupfireAgent({
+  memory: '1GB',
+  timeout: 100
+})